+++ src/main/java/org/fulib/patterns/PatternMatcher.java
@@ -37,6 +37,8 @@ public class PatternMatcher
 
    // =============== Properties ===============
 
+   // --------------- Deprecated ---------------
+
    /**
     * @return the map from pattern object to tables of matching values
     *
@@ -48,20 +50,11 @@ public class PatternMatcher
       return new LinkedHashMap<>(this.object2TableMap);
    }
 
-   /**
-    * @param pattern
-    *    the pattern object
-    *
-    * @return the table of matching values, or {@code null} if the pattern object was not reached
-    *
-    * @since 1.2
-    */
-   public ObjectTable getMatchTable(PatternObject pattern)
-   {
-      return this.object2TableMap.get(pattern);
-   }
+   // --------------- Debugging ---------------
 
    /**
+    * @return {@code true} if debug logging is enabled
+    *
     * @since 1.3
     */
    public boolean isDebugLogging()
@@ -70,6 +63,9 @@ public class PatternMatcher
    }
 
    /**
+    * @param eventLogging
+    *    determines whether debug logging is on ({@code true}) or off ({@code false})
+    *
     * @since 1.3
     */
    public void setDebugLogging(boolean eventLogging)
@@ -86,6 +82,9 @@ public class PatternMatcher
    }
 
    /**
+    * @return the list of debug events that happened during the match process.
+    * Empty if debug logging is turned off, or when {@link #match()} or its variants have not been called yet.
+    *
     * @since 1.3
     */
    public List<DebugEvent> getDebugEvents()
@@ -93,7 +92,11 @@ public class PatternMatcher
       return this.events == null ? Collections.emptyList() : Collections.unmodifiableList(this.events);
    }
 
+   // --------------- Roots ---------------
+
    /**
+    * @return the root pattern objects
+    *
     * @since 1.3
     */
    public List<PatternObject> getRootPatternObjects()
@@ -102,6 +105,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param patternObject
+    *    the root pattern object to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootPatternObjects(PatternObject patternObject)
@@ -111,6 +119,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param patternObjects
+    *    the root pattern objects to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootPatternObjects(PatternObject... patternObjects)
@@ -120,6 +133,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param patternObjects
+    *    the root pattern objects to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootPatternObjects(Collection<? extends PatternObject> patternObjects)
@@ -129,6 +147,8 @@ public class PatternMatcher
    }
 
    /**
+    * @return the root objects
+    *
     * @since 1.3
     */
    public List<Object> getRootObjects()
@@ -137,6 +157,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param object
+    *    the root object to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootObjects(Object object)
@@ -146,6 +171,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param objects
+    *    the root objects to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootObjects(Object... objects)
@@ -155,6 +185,11 @@ public class PatternMatcher
    }
 
    /**
+    * @param objects
+    *    the root objects to add
+    *
+    * @return this instance, to allow method chaining
+    *
     * @since 1.3
     */
    public PatternMatcher withRootObjects(Collection<?> objects)
@@ -165,13 +200,20 @@ public class PatternMatcher
 
    // =============== Methods ===============
 
+   // --------------- Matching Start ---------------
+
    /**
-    * Matches the pattern against the structure of objects that can be discovered from the start objects,
+    * Matches the pattern against the structure of objects that can be discovered from the root objects,
     * and returns the table of matched values corresponding to the named pattern object.
+    * <p>
+    * Equivalent to:
+    * <pre>{@code
+    *    this.match(pattern.getObject(rootPatternObjectName), rootObjects);
+    * }</pre>
     *
-    * @param patternObjectName
-    *    the name of the pattern object the results are requested for
-    * @param startObjects
+    * @param rootPatternObjectName
+    *    the name of the root pattern object, and for which the results are requested for
+    * @param rootObjects
     *    the root objects for object structure discovery
     *
     * @return the table of matched values corresponding to the named pattern object
@@ -180,18 +222,26 @@ public class PatternMatcher
     *    if there were unmatched constraints, possibly due to disconnected pattern objects
     * @see #match(PatternObject, Object...)
     */
-   public ObjectTable match(String patternObjectName, Object... startObjects)
+   public ObjectTable match(String rootPatternObjectName, Object... rootObjects)
    {
-      return this.match(this.pattern.getObject(patternObjectName), startObjects);
+      return this.match(this.pattern.getObject(rootPatternObjectName), rootObjects);
    }
 
    /**
-    * Matches the pattern against the structure of objects that can be discovered from the start objects,
+    * Matches the pattern against the structure of objects that can be discovered from the root objects,
     * and returns the table of matched values corresponding to the given pattern object.
+    * <p>
+    * Equivalent to:
+    * <pre>{@code
+    *    this.withRootPatternObjects(rootPatternObject);
+    *    this.withRootObjects(rootObjects);
+    *    this.match();
+    *    return this.getMatchTable(rootPatternObject);
+    * }</pre>
     *
-    * @param patternObject
-    *    the pattern object the results are requested for
-    * @param startObjects
+    * @param rootPatternObject
+    *    the root pattern object, and for which the results are requested for
+    * @param rootObjects
     *    the root objects for object structure discovery
     *
     * @return the table of matched values corresponding to the given pattern object
@@ -200,17 +250,21 @@ public class PatternMatcher
     *    if there were unmatched constraints, possibly due to disconnected pattern objects
     * @since 1.2
     */
-   public ObjectTable match(PatternObject patternObject, Object... startObjects)
+   public ObjectTable match(PatternObject rootPatternObject, Object... rootObjects)
    {
-      this.withRootPatternObjects(patternObject);
-      this.withRootObjects(startObjects);
+      this.withRootPatternObjects(rootPatternObject);
+      this.withRootObjects(rootObjects);
 
       this.match();
 
-      return this.getMatchTable(patternObject);
+      return this.getMatchTable(rootPatternObject);
    }
 
    /**
+    * Matches the pattern against the structure of objects that can be discovered from the root objects
+    *
+    * @throws NoApplicableConstraintException
+    *    if there were unmatched constraints, possibly due to disconnected pattern objects
     * @since 1.3
     */
    public void match()
@@ -267,32 +321,79 @@ public class PatternMatcher
       }
    }
 
+   // --------------- Result Extraction ---------------
+
    /**
+    * @param pattern
+    *    the pattern object
+    *
+    * @return the table of matching values, or {@code null} if the pattern object was not reached
+    *
+    * @since 1.2
+    */
+   public ObjectTable getMatchTable(PatternObject pattern)
+   {
+      return this.object2TableMap.get(pattern);
+   }
+
+   /**
+    * Attempts to retrieve exactly one unique match for the given pattern object.
+    * Must be called after {@link #match()} or its variants.
+    *
+    * @param <T>
+    *    the expected type of the result
+    * @param patternObject
+    *    the pattern object results are requested for
+    *
+    * @return the match for the given pattern object
+    *
+    * @throws NoMatchException
+    *    if no objects match
+    * @throws AmbiguousMatchException
+    *    if more than one distinct objects match
     * @since 1.3
     */
    public <T> T findOne(PatternObject patternObject)
    {
       final ObjectTable table = this.getMatchTable(patternObject);
       final int count = table.rowCount();
-      if (count == 1)
-      {
-         return (T) table.iterator().next();
-      }
       if (count == 0)
       {
          throw new NoMatchException(patternObject);
       }
-      throw new AmbiguousMatchException(patternObject, table.toList());
+      if (count == 1)
+      {
+         return (T) table.iterator().next();
+      }
+      // there may be the same object multiple times in the column
+      final Set<Object> set = table.toSet();
+      if (set.size() == 1)
+      {
+         return (T) set.iterator().next();
+      }
+      throw new AmbiguousMatchException(patternObject, set);
    }
 
    /**
+    * Retrieves the matches for the given pattern object.
+    * Must be called after {@link #match()} or its variants.
+    *
+    * @param <T>
+    *    the expected type of the results
+    * @param patternObject
+    *    the pattern object results are requested for
+    *
+    * @return the matches for the given pattern object. May be empty.
+    *
     * @since 1.3
     */
-   public <T> List<T> findAll(PatternObject patternObject)
+   public <T> Set<T> findAll(PatternObject patternObject)
    {
       final ObjectTable table = this.getMatchTable(patternObject);
-      return (List<T>) table.toList();
+      return (Set<T>) table.toSet();
    }
+
+   // --------------- Implementation ---------------
 
    private boolean checkAttributeConstraint(List<AttributeConstraint> attributeConstraints)
    {
@@ -324,17 +425,11 @@ public class PatternMatcher
    private boolean checkMatchConstraint(List<MatchConstraint> matchConstraints)
    {
       // find roles from done to not done
-      constraintLoop:
       for (MatchConstraint constraint : matchConstraints)
       {
-         for (PatternObject patternObject : constraint.getObjects())
+         if (!this.areAllPatternObjectsAvailable(constraint))
          {
-            ObjectTable srcTable = this.object2TableMap.get(patternObject);
-
-            if (srcTable == null)
-            {
-               continue constraintLoop; //=====================
-            }
+            continue;
          }
 
          final ObjectTable table = this.object2TableMap.get(constraint.getObjects().get(0));
@@ -350,6 +445,11 @@ public class PatternMatcher
       }
 
       return false;
+   }
+
+   private boolean areAllPatternObjectsAvailable(MatchConstraint constraint)
+   {
+      return constraint.getObjects().stream().allMatch(this.object2TableMap::containsKey);
    }
 
    private boolean checkHasLink(List<RoleObject> roles)